45 - chore: Renamed useFetchEventById file
